backend/requirements.txt
alembic==1.18.4
altgraph==0.17.4
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.3
narwhals==2.3.0
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfkit==1.0.0
pefile==2023.2.7
plotly==6.3.0
psycopg2-binary==2.9.11
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.11.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.46
trio==0.30.0
trio-websocket==0.12.2
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.6
wsproto==1.2.0